Execution implementation done
The child is able to execute the code, but it never returns it.
In the current implementation, no values are returned from the pool.
The children will not die until a something is inserted into the queue
and the pool is dead.

Some tests have been created, but they are far from exhaustive